backends/fastapi-backend/pyproject.toml
[tool.poetry]
name = "eventor-api"
version = "0.1.0"
description = "FastAPI instance to access the eventor service."
authors = ["synk <jairomer@protonmail.com>"]
license = "GNUv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.2.0"
fastapi = "^0.111.0"
pillow = "^10.3.0"
requests = "^2.31.0"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
pydantic = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
